StandaloneSmmPkg: Obtain SMRAM ranges from Hoblist
This patch changes the mechanism of obtaining the SMRAM ranges in the
StandaloneSmmPkg. The previous approach relied on a Guided HOB that
contains a copy of SMM_CORE_PRIVATE_DATA which in turn provides the
SMRAM ranges. This is not suitable when the SMM Foundation is being
initialised in the SEC phase by non UEFI/PI software. This patch
introduces the use of a SMRAM memory Hob as described in Volume 3: PI
Specification as an alternative to the previous approach. The header
that defines the relevant data structures has been imported from the
IntelFrameworkPkg.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>