StandaloneSmmPkg: Get MP information from Hoblist
Dynamic allocation and identification of per-CPU data structures
relies on knowledge of the number of CPUs and their IDs in the
system. This information is populated in a MP_INFORMATION_HOB_DATA by
the software entity responsible for initialising the MM
Foundation. This patch adds support for retrieving it from the HOB.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>