StandaloneSmmPkg: Add support for handling GUIDed events
This patch populates the CPU driver entry point with the logic to
handle GUIDed events.

Upon receipt of an event, it is determined whether the event ID
(passed by ARM Trusted Firmware) has a 1:1 mapping with a GUID or
not. In the former case, the GUID is copied into the per-CPU buffer
(PerCpuGuidedEventContext) by the CPU driver. In the latter case, it
is assumed that the event is a MM_COMMUNICATE SMC and the parameters
are present in X1 and X2 (X1 passed by Trusted Firmware. X2 is
deprecated). The communication buffer in X1 is expected to lie inside
the NS communication buffer allocated by ARM Trusted Firmware. It is
copied into the per-CPU buffer suitable for use in the secure world.

The MM Entry point is called which retrieves the communication buffer
using the GUIDed event management protocol. An attempt is made to
handle the event using the retrieved GUID. In case of an error, a
further attempt is made to handle the event using the traditional
mechanism of handling MM events.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>