StandaloneSmmPkg: Map in MM communication buffer during PEI
One of the arguments of the MM communication SMC is a pointer to a communication
buffer that contains the payload for the SMC. For simplicity and efficiency
reasons, it is desirable to prevent the secure world from mapping and unmapping
a potentially different buffer for each such SMC. Hence, the secure world
reserves a memory range for all MM communication. The extents of this memory
range are baked into the UEFI platform port. This patch adds support to obtain
these extents for the ARM Base FVP and map the memory region in UEFI memory
during the PEI phase.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>