StandaloneSmmPkg: Rename ARM CPU driver and functions
The PiMmCpuStandaloneMm CPU driver interfaces with third party (TP) firmware to
manage MM events. The TP firmware could be running in a privileged EL or
possibly in the same EL as the MM environment. The MM environment could be
running in an unprivileged or privileged EL. For e.g. the MM environment could
be running in S-EL0 and interfacing with ARM TF running in EL3 through a S-EL1
shim layer.

This patch renames the CPU driver and its internal functions to better reflect
the environment in which it executes.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>