StandaloneSmmPkg: Move ARM TF CPU driver under a closely named directory
The PiMmCpuTpFw directory contains the Standalone MM CPU driver
responsible for interfacing with ARM Trusted Firmware. This patch
changes the name of this directory and the INF file to better reflect
the contents that it holds. A subsequent patch will reorganise the
directory contents in a similar vein.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>