ArmPkg: Remove dead code & make MM Communication driver spec. compliant
This patch removes code to issue an SMC to obtain the location and size of the
MM communication buffer from the Secure world. Since it is not possible to add
memory regions to the EDK2 memory map at runtime this approach should be
revisited later. Also, the arguments of the MM_COMMUNICATE SMC are being
marshalled as per the MM specification.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>