StandaloneSmmPkg: Use Guided HOB to determine MM communication buffer
The AArch64 Entry point library produces a Hoblist that contains a
guided HOB with the address and size of the buffer meant for
communication with the Normal World and allocated by ARM Trusted
Firmware. The CPU driver was earlier relying on retrieving this
information from an SMRAM descriptor with a custom flag. This patch
updates the code to use the guided HOB to obtain this information.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>