StandaloneSmmPkg: Pass per-cpu shared buffer size to partition
This patch modifies the earlier approach of sending the total size of
the shared buffer between ARM TF and the EDK2 Standalone MM image as a
part of the boot information. The per-cpu size of the buffer is
populated instead now. The entry point code now uses the cpu count to
determine the total size of the shared buffer.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>